// apps/react-frontend/src/api/client/models/create-container-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ContainerLocation } from './container-location';

/**
 * 
 * @export
 * @interface CreateContainerRequest
 */
export interface CreateContainerRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateContainerRequest
     */
    'id': string;
    /**
     * 
     * @type {ContainerLocation}
     * @memberof CreateContainerRequest
     */
    'location'?: ContainerLocation;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerRequest
     */
    'notes'?: string;
}